LayoutTests/fast/forms/ios/form-control-refresh/select/choose-select-option.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
    <head>
        <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
        <script src="../../../../../resources/js-test.js"></script>
        <script src="../../../../../resources/ui-helper.js"></script>
    </head>
<body>
<select id="select">
    <option>January</option>
    <option>February</option>
    <option>March</option>
    <option>April</option>
    <option>May</option>
    <option>June</option>
    <option>July</option>
    <option>August</option>
    <option>September</option>
    <option>October</option>
    <option>November</option>
    <option>December</option>
</select>

<select id="grouped">
    <optgroup label="Winter">
        <option>January</option>
        <option>February</option>
        <option>March</option>
    </optgroup>
    <optgroup label="Spring">
        <option>April</option>
        <option>May</option>
        <option>June</option>
    </optgroup>
    <optgroup label="Summer">
        <option>July</option>
        <option>August</option>
        <option>September</option>
    </optgroup>
    <optgroup label="Fall">
        <option>October</option>
        <option>November</option>
        <option>December</option>
    </optgroup>
</select>
</body>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test verifies that tapping on a select element and choosing an option via the presented context menu updates the element's value.");

    shouldBeEqualToString("select.value", "January");
    await UIHelper.activateElement(select);
    await UIHelper.waitForContextMenuToShow();
    await UIHelper.selectFormAccessoryPickerRow(3);
    await UIHelper.waitForContextMenuToHide();
    shouldBeEqualToString("select.value", "April");

    shouldBeEqualToString("grouped.value", "January");
    await UIHelper.activateElement(grouped);
    await UIHelper.waitForContextMenuToShow();
    await UIHelper.selectFormAccessoryPickerRow(9);
    await UIHelper.waitForContextMenuToHide();
    shouldBeEqualToString("grouped.value", "October");

    finishJSTest();
});
</script>
</html>
